backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
PyMuPDF

# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface

# Embeddings + Vector storage
faiss-cpu
transformers
torch
